% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_dir_data.R
\name{change_dir_data}
\alias{change_dir_data}
\title{Change data directory}
\usage{
change_dir_data(newdirectory)
}
\arguments{
\item{newdirectory:}{the new directory from which to load the data}
}
\description{
The function changes the directory from where the data is loaded
}
\details{
One can download the data using download_iots() and then use this function to change the directory with this function.
Manual change of the data directory is also possible when loading data.
}
\examples{
\dontrun{change_dir_data("X://data")}
 
}
