% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient.R
\name{coefficient}
\alias{coefficient}
\title{Find the coefficient matrix of an International Input-Output Table}
\usage{
coefficient(iot)
}
\arguments{
\item{iot:}{List with elements in a Input-Output Table (load via [load_iot()])}
}
\value{
The coefficient matrix of the Input-Output Table.
}
\description{
Function calculates the coefficient matrix for 
a given International Input-Output table.
}
\examples{
iot <- load_iot("WIOD2013", 2000)
coef <- coefficient(iot)
   
}
