% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrycat.R
\name{countrycat}
\alias{countrycat}
\title{Integer classification of countries}
\usage{
countrycat(country_lists, iots)
}
\arguments{
\item{country_lists:}{List of vectors with countries that belong to a certain category.}

\item{iots:}{Timeseries list of lists with elements of input-output table.}
}
\value{
vector with a classification of countries in integers.
}
\description{
Function generates a vector with a classification 
of countries in integers
}
\examples{
iots <- load_iots("WIOD2013", 2000:2001)
primary <- c("AtB", "C") 
transport <- c("60", "61", "62", "63")
industries <- industrycat(list(primary, transport), iots)
}
