% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryindustry_descr.R
\name{countryindustry_descr}
\alias{countryindustry_descr}
\title{Country - Industry description}
\usage{
countryindustry_descr(iot, countrycat = "None", industrycat = "None")
}
\arguments{
\item{iot:}{List with elements of input-output table.}

\item{countrycat:}{Vector with categorisation of countries (in integers).}

\item{industrycat:}{vector with categorisation of industries (in integers).}
}
\value{
Array with the country-industry description of the data.
}
\description{
Function gives the common country-industry description to the data
}
\details{
Mostly for use within functions to provide the country-industry description 
with its output.
}
\examples{
iot <- load_iot("WIOD2013", 2000)
descr <- countryindustry_descr(iot)

iots <- load_iots("WIOD2013", 2000:2001)
NAFTA <- c("USA", "MEX", "CAN")
BENELUX <- c("BEL", "NLD", "LUX") 
regions <- countrycat(list(NAFTA, BENELUX), iots)
primary <- c("AtB", "C") 
transport <- c("60", "61", "62", "63")
industries <- industrycat(list(primary, transport), iots)
descr_cat <- countryindustry_descr(iot, countrycat = regions, industrycat = industries) 

}
