% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_iot.R
\name{download_iot}
\alias{download_iot}
\title{Download Single IOT}
\usage{
download_iot(
  version,
  year,
  dir_data = get("dir_data", envir = paramEnv),
  dir_data_online = get("dir_data_online", envir = paramEnv)
)
}
\arguments{
\item{version:}{The international input-output database to be used}

\item{year:}{The year for which to add supplemental data}

\item{dir_data:}{The path towards the local directory in which to store the data. Most often using change_dir_data is preferred.}

\item{dir_data_online:}{the path towards the internet source with data. Normally no need to change this.}
}
\description{
Download a single IOT from the internet to your computer.
download_iots will download a full dataset of IOTs.
}
\examples{
\dontrun{
Using change_dir_data
change_dir_data("D:/Data")
download_iot("WIOD2013", 2000)

Using dir_data param
download_iot("WIOD2013", 2000, "D:/Data")
}
  
}
