% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_iots.R
\name{download_iots}
\alias{download_iots}
\title{Download Timeseries of IOTs}
\usage{
download_iots(
  version,
  dir_data = get("dir_data", envir = paramEnv),
  dir_data_online = get("dir_data_online", envir = paramEnv)
)
}
\arguments{
\item{version:}{The international input-output database to be used}

\item{dir_data:}{The path towards the local directory in which to store the data. Most often using change_dir_data is preferred.}

\item{dir_data_online:}{the path towards the internet source with data. Normally no need to change this.}
}
\description{
Download a timeseries IOT from the internet to your computer.
}
\examples{
\dontrun{
Using change_dir_data
change_dir_data("D:/Data")
download_iots("WIOD2013")

Using dir_data param
download_iots("WIOD2013", "D:/Data")
}
  
}
