% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_csv.R
\name{export_csv}
\alias{export_csv}
\title{Make a CSV file from measures in IOTs}
\usage{
export_csv(measures, iots, long = FALSE, filename = "choose")
}
\arguments{
\item{measures:}{vector with the names of the measures}

\item{iots:}{list of input-output tables for which measures are already calculated}

\item{long:}{Whether data is in long or wide format.}

\item{filename:}{Where to store the data}
}
\description{
Function to extract measures from IOTs into a CSV file
}
\details{
filename = "choose" is default and will prompt a file-choose dialog.

First element in vector will be used to find the description of the data
Make sure the output of the measures are of the same length. 
(i.e. all need to be on e.g. country-industry, country or industry level)
}
\examples{
\dontrun{
iots <- load_iots("WIOD2013", 2000:2001)
iots <- oniots(wiod_gii, iots)

Not specifying directory prompts a file-choose dialog
export_csv("gii", iots)
export_csv("gii", iots, long = TRUE)

Save table in working directory
export_csv("gii", iots, filename = "myresults.csv")

Or specify a directory:
export_csv("gii", iots, filename = "D:/Research/myresults.csv")
}

}
\seealso{
export_dataframe(): [export_dataframe()]
}
