% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_dataframe.R
\name{export_dataframe}
\alias{export_dataframe}
\title{Make a dataframe from measures in IOTs}
\usage{
export_dataframe(measures, iots, long = FALSE)
}
\arguments{
\item{measures:}{vector with the names of the measures.}

\item{iots:}{list of input-output tables for which measures are already calculated.}

\item{long:}{Whether data is in long or wide format.}
}
\description{
Function to extract measures from IOTs into a dataframe.
}
\details{
First element in vector will be used to find the description of the data
Make sure the output of the measures are of the same length. 
(i.e. all need to be on e.g. country-industry, country or industry level)
}
\examples{
iots <- load_iots("WIOD2013", 2000:2001)
iots <- on_iots(gii, iots)
df_wide <- export_dataframe("gii", iots)
df_long <- export_dataframe("gii", iots, long = TRUE)

}
