% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_total.R
\name{fd_total}
\alias{fd_total}
\title{Total Final Demand}
\usage{
fd_total(iot, regions = "None", industries = "None")
}
\arguments{
\item{iot:}{List with elements in a Input-Output Table (load via [load_iot()])}
}
\value{
List with elements of Input-Output Table and added the total final demand (fd) 
and a country-industry categorization (fd_descr)
}
\description{
Function returns the aggregated total final demand at the country-industry level
}
\examples{
iot <- load_iot("WIOD2013", 2000)
iot <- fd_total(iot)

iots <- load_iots("WIOD2013", 2000:2001)
iots <- on_iots(fd_total, iots)
iots <- on_iots(gii, iots)
df_fs <- export_dataframe(c("fd_total", "gii"), iots)
   
}
