% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvas.R
\name{fvas}
\alias{fvas}
\title{Find Foreign and Regional Value Added Share}
\usage{
fvas(iot, regions = "None", industries = "None")
}
\arguments{
\item{iot:}{List with elements in a Input-Output Table (load via [load_iot()])}

\item{regions:}{Vector with categorisation of countries.}

\item{industries:}{Vector with categorisation of industries.}
}
\value{
The input-output table list with 
the Foreign Value Added Share measure (fvas),
the Regional Foreign Value Added Share measure (rfvas), 
and the Global Foreign Value Added Share measure (gfvas)
and a description of the countries and industries 
(fvas_descr, rfvas_descr, gfvas_des) for use in [export_dataframe()].
}
\description{
Calculates the foreign value added and regional value added shares as in 
Bart Los, Marcel P Timmer, Gaaitzen J de Vries
How global are global value chains? A new approach to measure international fragmentation#'
}
\details{
regions defines the regions that are used to calculate RFVAS. 
If none selected, RFVAS will be equal to FVAS.

Full reference:
Bart Los, Marcel P Timmer, Gaaitzen J de Vries, 2015,
How global are global value chains? A new approach to measure international fragmentation
Journal of Regional Science, volume 55, issue 1, pages 66-92
}
\examples{
iot <- load_iot("WIOD2013", 2000)
iot <- fvas(iot)

iots <- load_iots("WIOD2013", 2000:2001)
NAFTA <- c("USA", "MEX", "CAN")
BENELUX <- c("BEL", "NLD", "LUX") 
regions <- countrycat(list(NAFTA, BENELUX), iots)
primary <- c("AtB", "C") 
transport <- c("60", "61", "62", "63")
industries <- industrycat(list(primary, transport), iots)

iots <- on_iots(fvas, iots, regions = regions, industries = industries)
df_gii <- export_dataframe(c("fvas", "rfvas", "gfvas"), iots)

}
