% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/industrycat.R
\name{industrycat}
\alias{industrycat}
\title{Integer classification of Industries}
\usage{
industrycat(industry_lists, iots)
}
\arguments{
\item{industry_lists:}{List of vectors with industries that belong to a certain category.}

\item{iots:}{Timeseries list of lists with elements of input-output table.}
}
\value{
vector with a classification of countries in integers.
}
\description{
Function generates a vector with a classification 
of industries in integers
}
\examples{
iots <- load_iots("WIOD2013", 2000:2001)
NAFTA <- c("USA", "MEX", "CAN")
BENELUX <- c("BEL", "NLD", "LUX") 
regions <- countrycat(list(NAFTA, BENELUX), iots)
}
