% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontief.R
\name{leontief}
\alias{leontief}
\title{Leontief Inverse of an International Input-Output Table}
\usage{
leontief(iot)
}
\arguments{
\item{iot:}{List with elements in a Input-Output Table (load via [load_iot()])}
}
\value{
List with elements of Input-Output Table and the Leontief inverse.
}
\description{
Function calculates the Leontief inverse for 
a given International Input-Output table.
}
\examples{
iot <- load_iot("WIOD2013", 2000)
iot <- leontief(iot)
   
}
