% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_extra_iots.R
\name{load_extra_iots}
\alias{load_extra_iots}
\title{Load extra data to multiple IOTs.}
\usage{
load_extra_iots(iots, extra, directory = get("dir_data", envir = paramEnv))
}
\arguments{
\item{iots:}{List of Input-Output Tables (load via [load_iots()])}

\item{extra:}{Name of the extra data that is required.}

\item{directory:}{directory to where the input-output tables are stored.}
}
\description{
Function to load extra data to a single internation Input-Output Table (IOT).
}
\details{
Directory only needs to direct to the main storage directory, right version and year will be taken care of. 
Default is loading via the internet. Users who use IOTs more often are advised to make a local copy.
Local copy of online data can be made via download_iots()
}
\examples{
\dontrun{Using online data:}
iots <- load_iots("WIOD2013")
iots <- load_extra_iots(iots, "SEA")

iots <- load_iots("WIOD2013", 2000:2001)
iots <- load_extra_iots(iots, "SEA")

\dontrun{Using local data:
iots <- load_iots("WIOD2013", directory = "D:/Data")
iots <- load_extra_iots(iots, "SEA", directory = "D:/Data")
}

 
}
