% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_iot.R
\name{load_iot}
\alias{load_iot}
\title{Load a single International Input-Output Table.}
\usage{
load_iot(version, year, directory = get("dir_data", envir = paramEnv))
}
\arguments{
\item{version:}{version of IOT}

\item{year:}{year of IOT}

\item{directory:}{directory to where the input-output tables are stored.}
}
\description{
Function loads a single internation Input-Output Table (IOT).
}
\details{
Versions currently supported are WIOD2013 (World Input Output Database, version 2013), 
WIOD2016 (World Input Output Database, version 2013) 
and ICIOT2018 (Inter-Country Input Output Tables, version 2018).

To load a time-series of international input-output tables use load_iots()

Directory only needs to direct to the main storage directory, right version and year will be taken care of. 
Default is loading via the internet. Users who use IOTs more often are advised to make a local copy.
Local copy of online data can be made via download_iots()
}
\examples{
\dontrun{Using online data:}
iot <- load_iot("WIOD2013", 2000)

\dontrun{Using local data:
iot <- load_iot("WIOD2013", 2000, directory = "D:/Data")}

 
}
