% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_iots.R
\name{load_iots}
\alias{load_iots}
\title{Load a time-series of International Input-Output Table.}
\usage{
load_iots(version_database, years = NULL, directory = get("dir_data",
  envir = paramEnv))
}
\arguments{
\item{version:}{version of IOTs}

\item{year:}{years for which you want the IOT. Default is all years.}

\item{directory:}{directory to where the input-output tables are stored.}
}
\description{
Function loads a time-series of internation Input-Output Tables (IOTs).
}
\details{
Versions currently supported are WIOD2013 (World Input Output Database, version 2013), 
WIOD2016 (World Input Output Database, version 2013) 
and ICIOT2018 (Inter-Country Input Output Tables, version 2018).

Directory only needs to direct to the main storage directory, right version and year will be taken care of. 
Default is loading via the internet. Users who use IOTs more often are advised to make a local copy.
Local copy of online data can be made via download_iots()
}
\examples{
\dontrun{Using online data:}
iots <- load_iots("WIOD2013")

\dontrun{Using online data, only some years:}
iots <- load_iots("WIOD2013", 2000:2002)

\dontrun{Using local data:
iots <- load_iots("WIOD2013", directory = "D:/Data")}

}
