% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_iots.R
\name{on_iots}
\alias{on_iots}
\title{Run over a list of International Input-Output Tables (IOTs)}
\usage{
on_iots(fun, iots, ...)
}
\arguments{
\item{fun:}{The function to run over.}

\item{iots:}{the list of IOTs}

\item{...:}{Any other arguments needed for function in fun,}
}
\value{
List of international input-output tables with the output of the 
function call added to each IOT.
}
\description{
[oniots()] executes the same function 
on a list of international input-output tables (IOTs)
}
\details{
fun should be a function that appends its output to the IOT.
}
\examples{
iots <- load_iots("WIOD2013", 2000:2001)
iots <- on_iots(gii, iots)
 
\dontrun{Example with extra arguments}
NAFTA <- c("USA", "MEX", "CAN")
BENELUX <- c("BEL", "LUX", "NLD") 
regions <- countrycat(list(NAFTA, BENELUX), iots)
primary <- c("AtB", "C") 
transport <- c("60", "61", "62", "63")
industries <- industrycat(list(primary, transport), iots)

iots <- on_iots(gii, iots, regions = regions, industries = industries)
df_gii <- export_dataframe("gii", iots)

}
