% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_dir_data.R
\name{reset_dir_data}
\alias{reset_dir_data}
\title{Reset data directory to online}
\usage{
reset_dir_data()
}
\description{
The function changes the directory from where the data is loaded back to the online data repository
}
