% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vax_functional.R
\name{vax_functional}
\alias{vax_functional}
\title{VAX-Domestic for Functional Specialization}
\usage{
vax_functional(iot)
}
\arguments{
\item{iot:}{List with elements in a Input-Output Table (load via [load_iot()])}
}
\description{
The Domestic Value Added Export that is used in the measure of functional specialization in
Timmer,Miroudot and De Vries (2018), Functional specialisation in trade.
}
\details{
Used [func_spec]

Full Reference: Marcel Timmer, Sébastien Miroudot and Gaaitzen de Vries (2018), 
Functional specialisation in trade., 
Journal of Economic Geography, Volume 19, Issue 1, Pages 1–30
}
\examples{
iot <- load_iot("WIOD2013", 2000)
iot <- vax_functional(iot)

}
