% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_dir_data.R
\name{change_dir_data}
\alias{change_dir_data}
\title{Change data directory}
\usage{
change_dir_data(newdirectory)
}
\arguments{
\item{newdirectory:}{the new directory from which to load the data and where to store data}
}
\description{
The function changes the directory from where the data is loaded
}
\details{
Use this function to change the directory from where to load downloaded IOTs with load_iot or load_iots. 
If a requested IOT is not yet downloaded, load_iot or load_iots will store the data first in this directory
and load the newly downloaded local file.
}
\examples{
\dontrun{change_dir_data("X://data")}
 
}
