% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_iot.R
\name{download_iot}
\alias{download_iot}
\title{Download Single IOT}
\usage{
download_iot(version, year, dir_data = get("dir_data", envir = paramEnv),
  dir_data_online = get("dir_data_online", envir = paramEnv))
}
\arguments{
\item{dir_data:}{The path towards the local directory in which to store the data.}

\item{dir_data_online:}{the path towards the internet source with data.Normally no need to change.}
}
\description{
Download a single IOT from the internet to your computer
}
