% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_iots.R
\name{load_iots}
\alias{load_iots}
\title{Load a time-series of International Input-Output Table.}
\usage{
load_iots(version_database, years = NULL, directory = get("dir_data",
  envir = paramEnv))
}
\arguments{
\item{version:}{version of IOTs}

\item{year:}{years for which you want the IOT. Default is all years.}

\item{directory:}{directory to where the input-output tables are stored. Most often preferable to change using change_dir_data().}
}
\description{
Function loads a time-series of internation Input-Output Tables (IOTs).
}
\details{
Versions currently supported are WIOD2013 (World Input Output Database, version 2013), 
WIOD2016 (World Input Output Database, version 2013) 
and ICIOT (Inter-Country Input Output Tables, version 2018).

If directory is changed using change_dir_data, the function will download the requested data prior to loading.
One can also download data using download_iots.
}
\examples{
\dontrun{Using online data:}
iots <- load_iots("WIOD2013")

\dontrun{Using online data, only some years:}
iots <- load_iots("WIOD2013", 2000:2002)

\dontrun{Using local data:
change_dir_data("D:/Data")
iots <- load_iots("WIOD2013")}

}
